Grid cutter function implemented
The function makes the radial cut, properly (even though the zeta of the grid and the Psi_p are not the same (check how they relate)), and the poloidal cut is also added, although not sure how to check it graphically. It needs to be checked. The average is still done in flux surfaces probably further than the Xgrid range. It is needed to check. But the cutter works in general terms. The cut is applied on volX2d (line 347 of feltordiag.cu).