Code with partial fsa and total fsa
With this update, the code already has the original fsa and a partial fsa with a certain range. Still, the scaling from the average is not certain where it is supposed to be done, and the average close to the X-point and 0 Psi_p kind of diverges, so I need to check that how it is done, as it is not being done properly.