Convolution adaptation
There are new coordinates (eta, psi2) for the X-point grid for the convolution function. The map is also implemented in xcc and ycc variables in this coordinates (eta,psi2). Also the convolution variables are in this parameters. It is still to be done: The time coordinate (not sure how it works yet) and introducing the values to the convoluted variables.