Convolution update
I have achieved to create only the new coordinates eta, psi2, with xcc, ycc and all the convolution variables inside them. Not sure about time, as when I try it to make it advance, Paraview crashes, but the matrices look to change reasonably good. Now, what I would probably need to do is to maybe cut the grid or make it in some way that the visualization does not saturates with the region with eta>0, work on the transformation and check if the convolution function works relativelly good.