2D convolution almost implemented and starting 1d
The 2D convolution is too time consuming as it is defined right now, going in all the points of the 2D grid. It will be neccesary to rethink the way of defining it. for the 1D, there is already a single eta coordinate with the variables, but I have to check how to introduce them in the variables.