Convolution 1D working
The Convolution in the LCFS works (although it takes a long time, around 40 seconds per time iteration)). It will be neccesary to check if parallelizing it can go faster.It will be also neccesary to introduce the poloidal angle variable in the eta coordinate to plot it in the x axis. Check also how to improve the oscillations of the blobs, that do not disappear completely.