Commented code
I have added descriptions in the Raul_convolution.h and Raul_grid_cutter.h and comments on these two documents and the feltordiag.cu on what has been changed mainly. As it is right now, the convolution function makes the average of two different angles(7.5 and 15), as it seemed to work in a better way, but this is still waiting for a discussion. Many thins could be taken out of the do_compute loops in both functors files, but no idea how to make that.