FINAL VERSION OF NEW DIAGNOSTIC
The code is already described and clean to be pushed. It makes the partial flux surface average and the convolution as it is defined in feltor.pdf. It is checked that the flux surface average is conserved in the convolution. It also has an implemented cut function to obtain radial cuts on any variable to check its poloidal dependence at any radial position.